feat: 테스트코드에 작성한 내용 계산기 클래스로 분리
